src/streaming/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --output-file requirements.txt
avro-python3==1.10.2
    # via streaming
kafka-python==2.2.13
    # via streaming
numpy==2.3.1
    # via opencv-python
opencv-python==4.11.0.86
    # via streaming
python-dotenv==1.1.1
    # via streaming
websockets==15.0.1
    # via streaming
